requirements.txt
rembg==2.0.50
Pillow==10.0.0
numpy==1.24.3
onnxruntime==1.15.1
torch==2.0.1
torchvision==0.15.2 